# Trim examples ahead of time for auto packing