--- a/test.yaml
+++ b/test.yaml
@@ -1,0 +1,99 @@
+
+seed: 17
+model:
+  name: hf_causal_lm
+  use_mpt: false
+  pretrained: true
+  init_device: mixed
+  use_auth_token: true
+  config_overrides:
+    attn_config:
+      attn_impl: triton
+      attn_uses_sequence_id: false
+    max_seq_len: 4096
+  use_train_metrics: true
+  attention_patch_type: null
+  use_flash_attention_2: false
+  pretrained_model_name_or_path: mosaicml/mpt-7b
+callbacks:
+  lr_monitor: {}
+  scheduled_gc:
+    batch_interval: 1000
+  speed_monitor:
+    window_size: 1
+  memory_monitor: {}
+load_path: null
+optimizer:
+  lr: 1.0e-06
+  name: decoupled_lionw
+  betas:
+  - 0.9
+  - 0.95
+  weight_decay: 1.0e-06
+precision: amp_bf16
+scheduler:
+  name: cosine_with_warmup
+  alpha_f: 0
+  t_warmup: 0.02dur
+tokenizer:
+  name: mosaicml/mpt-7b
+  kwargs:
+    revision: main
+    model_max_length: 4096
+algorithms:
+  gradient_clipping:
+    clipping_type: norm
+    clipping_threshold: 1
+autoresume: true
+code_paths: []
+eval_first: false
+log_config: true
+fsdp_config:
+  verbose: true
+  mixed_precision: PURE
+  state_dict_type: sharded
+  use_orig_params: true
+  forward_prefetch: false
+  backward_prefetch: BACKWARD_POST
+  limit_all_gathers: true
+  sharding_strategy: FULL_SHARD
+  sync_module_states: false
+  activation_cpu_offload: false
+  activation_checkpointing: true
+  load_monolith_rank0_only: false
+  activation_checkpointing_reentrant: false
+icl_seq_len: 1024
+max_seq_len: 4096
+dist_timeout: 600
+max_duration: 30ba
+progress_bar: false
+train_loader:
+  name: finetuning
+  dataset:
+    split: train
+    hf_name: mosaicml/dolly_hhrlhf
+    shuffle: true
+    max_seq_len: 4096
+    eos_token_id: 0
+    packing_ratio: auto
+    allow_pad_trimming: false
+    decoder_only_format: true
+  drop_last: true
+  pin_memory: true
+  num_workers: 8
+  prefetch_factor: 2
+  persistent_workers: true
+eval_interval: 1
+log_to_console: true
+save_overwrite: false
+load_ignore_keys: null
+python_log_level: debug
+save_ignore_keys: null
+load_weights_only: false
+save_weights_only: false
+console_log_interval: 1ba
+device_eval_batch_size: 12
+eval_subset_num_batches: 20
+global_train_batch_size: 16
+device_train_microbatch_size: 12
+save_num_checkpoints_to_keep: 1