Previously during packing profiling, we were repeatedly trimming the same examples. The longer the trimmed examples are, the more expensive the operation, so for datasets with longer examples (like `sam-mosaic/iv4-chatml-4k`), we saw very noticeable performance impacts.

## Improved Profiling Times
Tested with this [yaml](https://gist.github.com/irenedea/eb646e0b7cb89a380f63ba0ccabf2aa5), changing only the dataset
`mosaicml/dolly_hhrlhf`
* before: 29.977840185165405 s
* after: 23.40160083770752 s

`sam-mosaic/iv4-chatml-4k`
* before: 131.2125198841095 s
* after: 45.155487060546875 s


## Tests
* Manually tested with `mosaicml/dolly_hhrlhf` and `sam-mosaic/iv4-chatml-4k`. Checked that the same packing ratio was selected